compiling but test hangs
