+++ src/Network/Ethereum/Web3/Contract.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DataKinds             #-}
+{-# LANGUAGE RecordWildCards             #-}
 {-# LANGUAGE FlexibleContexts      #-}
 {-# LANGUAGE FlexibleInstances     #-}
 {-# LANGUAGE GADTs                 #-}
@@ -47,12 +48,18 @@ module Network.Ethereum.Web3.Contract (
   ) where
 
 import           Control.Concurrent                     (threadDelay)
-import           Control.Concurrent.Async               (Async)
+import           Control.Concurrent.Async                     (Async)
 import           Control.Exception                      (throwIO)
-import           Control.Monad                          (forM, when)
+import           Control.Monad                          (forM, when, void)
 import           Control.Monad.IO.Class                 (liftIO)
+import           Control.Monad.Trans.Class      (lift)
+import           Control.Monad.Trans.Maybe      (MaybeT (..))
 import           Control.Monad.Trans.Reader             (ReaderT (..))
 import           Data.Maybe                             (listToMaybe, mapMaybe)
+import           Data.Machine                   (Is (Refl), ProcessT,
+                                                 Step (Await, Stop, Yield),
+                                                 autoT, runMachineT)
+import           Data.Machine.MealyT            (MealyT (..), arrM)
 import           Data.Monoid                            ((<>))
 import           Data.Proxy                             (Proxy (..))
 import qualified Data.Text                              as T
@@ -82,42 +89,156 @@ class Event e where
     -- | Event filter structure used by low-level subscription methods
     eventFilter :: Proxy e -> Address -> Filter
 
--- | 'event' spawns an asynchronous event filter to monitor the latest events
--- | logged by the contract at the given address for a particular event type. All
--- | events of type 'e' are composed of an indexed component 'i', and a
--- | non-indexed component 'ni'.
-event :: forall p e i ni.
-          ( Provider p
-         , Event e
-         , DecodeEvent i ni e
-         )
-       => Address
-       -> (e -> ReaderT Change (Web3 p) EventAction)
-       -> Web3 p (Async ())
-event a f = do
-    fid <- Eth.newFilter (eventFilter (Proxy :: Proxy e) a)
-    forkWeb3 $
-        let loop = do liftIO (threadDelay 1000000)
-                      changes <- Eth.getFilterChanges fid
-                      acts <- forM (mapMaybe pairChange changes) $ \(changeEvent, changeWithMeta) ->
-                        runReaderT (f changeEvent) changeWithMeta
-                      when (TerminateEvent `notElem` acts) loop
-        in do loop
-              Eth.uninstallFilter fid
-              return ()
-  where
-    prepareTopics = fmap (T.drop 2) . drop 1
-    pairChange :: DecodeEvent i ni e => Change -> Maybe (e, Change)
-    pairChange changeWithMeta = do
-      changeEvent <- decodeEvent changeWithMeta
-      return (changeEvent, changeWithMeta)
+data FilterChange a = FilterChange { filterChangeRawChange :: Change
+                                   , filterChangeEvent     :: a
+                                   }
 
+-- the halting case is provided by Nothing
+type HaltingMealyT m = MealyT (MaybeT m)
+
+reduceEventStream :: forall m s a . Monad m
+                  => HaltingMealyT m s [FilterChange a]
+                  -> (a -> ReaderT Change m EventAction)
+                  -> s
+                  -> m (Maybe s)
+reduceEventStream machine handler initialState =
+  let process = autoT machine
+  in go process
+
+  where go :: ProcessT (MaybeT m) s [FilterChange a] -> m (Maybe s)
+        go process = do
+          res <- runMaybeT $ runMachineT process
+          case res of
+            Nothing   -> return $ Just initialState
+            Just Stop -> return $ Just initialState -- Pretty sure this will never get called,
+                                                    -- due to autoT for MealyT
+            Just (Yield changes next) -> do
+              acts <- processChanges handler changes
+              if TerminateEvent `notElem` acts
+              then go next
+              else return Nothing
+            Just (Await f Refl next) -> go (f initialState) >> go next -- This is a guess
+
+
+        processChanges :: (a -> ReaderT Change m EventAction) -> [FilterChange a] -> m [EventAction]
+        processChanges handler changes = forM changes $ \FilterChange{..} ->
+                                           flip runReaderT filterChangeRawChange $
+                                                handler filterChangeEvent
+
+data FilterStreamState = FilterStreamState { fssCurrentBlock  :: BlockNumber
+                                           , fssInitialFilter :: Filter
+                                           , fssWindowSize    :: Integer
+                                           }
+
+
+-- | 'playLogs' streams the 'filterStream' and calls eth_getLogs on these
+-- | 'Filter' objects.
+playLogs :: forall p i ni e.
+            ( Provider p
+            , DecodeEvent i ni e
+            , Event e
+            )
+         => HaltingMealyT (Web3 p) FilterStreamState [FilterChange e]
+playLogs  = do
+  filter <- filterStream
+  changes <- arrM . const . lift $ Eth.getLogs filter
+  return $ mkFilterChanges changes
+
+pollFilter :: forall p i ni e s .
+              ( Provider p
+              , DecodeEvent i ni e
+              , Event e
+              )
+           => FilterId
+           -> DefaultBlock
+           -> HaltingMealyT (Web3 p) s [FilterChange e]
+pollFilter fid stop = MealyT $ \s -> do
+  bn <- lift $ Eth.blockNumber
+  if BlockWithNumber bn > stop
+  then do
+    lift $ Eth.uninstallFilter fid
+    MaybeT $ return Nothing
+  else do
+    lift . Web3 $ threadDelay 1000
+    changes <- lift $ Eth.getFilterChanges fid
+    return (mkFilterChanges changes, pollFilter fid stop)
+
+mkFilterChanges :: forall i ni e.
+                   ( Event e
+                   , DecodeEvent i ni e
+                   )
+                => [Change]
+                -> [FilterChange e]
+mkFilterChanges cs =
+  flip mapMaybe cs $ \c@Change{..} -> do
+    x <- decodeEvent c
+    return $ FilterChange c x
+
+filterStream :: forall p . Provider p
+             => HaltingMealyT (Web3 p) FilterStreamState Filter
+filterStream = MealyT $ \FilterStreamState{..} -> do
+  end <- lift . mkBlockNumber $ filterToBlock fssInitialFilter
+  if fssCurrentBlock > end
+  then MaybeT $ return Nothing -- halt
+  else return $ let to' = newTo end fssCurrentBlock fssWindowSize
+                    filter' = fssInitialFilter { filterFromBlock = BlockWithNumber fssCurrentBlock
+                                               , filterToBlock = BlockWithNumber to'
+                                               }
+                    MealyT next = filterStream
+                in (filter', MealyT $ \s -> next s { fssCurrentBlock = succ to' })
+
+      where succ :: BlockNumber -> BlockNumber
+            succ (BlockNumber bn) = BlockNumber $ bn + 1
+
+            newTo :: BlockNumber -> BlockNumber -> Integer -> BlockNumber
+            newTo upper (BlockNumber current) window = min upper . BlockNumber $ current + window
+
+
+event' :: forall p i ni e .
+          ( Provider p
+          , DecodeEvent i ni e
+          , Event e
+          )
+       => Filter
+       -> Integer
+       -> (e -> ReaderT Change (Web3 p) EventAction)
+       -> Web3 p ()
+event' fltr window handler = do
+  start <- mkBlockNumber $ filterFromBlock fltr
+  let initState = FilterStreamState { fssCurrentBlock = start
+                                    , fssInitialFilter = fltr
+                                    , fssWindowSize = window
+                                    }
+  mLastProcessedFilterState <- reduceEventStream playLogs handler initState
+  case mLastProcessedFilterState of
+    Nothing -> return ()
+    Just lastProcessedFilterState -> do
+      let BlockNumber lastBlock = fssCurrentBlock lastProcessedFilterState
+          pollingFromBlock = BlockNumber $ lastBlock + 1
+          pollTo = filterToBlock fltr
+      filterId <- Eth.newFilter fltr { filterFromBlock = BlockWithNumber pollingFromBlock }
+      void $ reduceEventStream (pollFilter filterId pollTo) handler ()
+
+event :: forall p i ni e .
+         ( Provider p
+         , DecodeEvent i ni e
+         , Event e
+         )
+      => Filter
+      -> (e -> ReaderT Change (Web3 p) EventAction)
+      -> Web3 p (Async ())
+event fltr handler = forkWeb3 $ event' fltr 0 handler
+
+mkBlockNumber :: Provider p => DefaultBlock -> Web3 p BlockNumber
+mkBlockNumber bm = case bm of
+  BlockWithNumber bn -> return bn
+  Earliest -> return 0
+  _ -> Eth.blockNumber
 
 class Method a where
-  -- | selector is used to compute the function selector for a given function type, defined as
-  -- | the hex string representation of the first 4 bytes of the hash of the signature.
   selector :: Proxy a -> T.Text
 
+-- | 'sendTx' is used to submit a state changing transaction.
 sendTx :: ( Generic a
           , GenericABIEncode (Rep a)
           , Provider p
@@ -132,6 +253,8 @@ sendTx call (dat :: a) =
   let sel = selector (Proxy :: Proxy a)
   in Eth.sendTransaction (call { callData = Just $ sel <> genericToData dat })
 
+-- | 'call' is used to call contract methods that have no state changing effects,
+-- | or to call m
 call :: ( Generic a
         , GenericABIEncode (Rep a)
         , Generic b
