+++ web3.cabal
@@ -46,7 +46,6 @@ library
                      , Network.Ethereum.Web3.Provider
                      , Network.Ethereum.Web3.Encoding
                      , Network.Ethereum.Web3.Contract
-                     , Network.Ethereum.Web3.Contract.Streaming
                      , Network.Ethereum.Web3.Encoding.Bytes
                      , Network.Ethereum.Web3.Encoding.Event
                      , Network.Ethereum.Web3.Encoding.Event.Internal
